% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addedLastTest.R
\name{addedLastTest}
\alias{addedLastTest}
\title{Compute weighted burden test}
\usage{
addedLastTest(
  wgt,
  snps,
  sumStats,
  snpAnnot = NULL,
  beta,
  se,
  chr,
  pos,
  ref,
  pval,
  R2cutoff,
  locChrom
)
}
\arguments{
\item{wgt}{character, name of expression model .RData file}

\item{snps}{data.frame, reference panel SNPs}

\item{sumStats}{data frame, GWAS summary statistics}

\item{snpAnnot}{data.frame, SNP REF/ALT annotations}

\item{beta}{character, colnames in sumStats that keeps the effect sizes}

\item{se}{character, colnames in sumStats that keeps the standard errors}

\item{chr}{character, colnames in sumStats that keeps the chromosome}

\item{pos}{character, colnames in sumStats that keeps the position}

\item{ref}{character, colnames in sumStats that keeps the reference allele}

\item{R2cutoff}{numeric, predictive R2 cutoff}

\item{locChrom}{character, chromosome of gene of interest}
}
\value{
Z-score and P-value of added last test
}
\description{
The function takes in a gene expression model in MOSTWAS form
and GWAS summary statistics and
carries out the weighted burden Z-test for a trait
}
