% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjR2.R
\name{adjR2}
\alias{adjR2}
\title{Compute McNemar's adjusted R2}
\usage{
adjR2(x, y, method = "pearson")
}
\arguments{
\item{x}{numeric vector}

\item{y}{numeric vector}

\item{method}{character, correlation method}
}
\value{
adjusted R2 of x and y
}
\description{
The function computes the adjusted R2 between two numeric vectors
}
