% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplifyTrans.R
\name{amplifyTrans}
\alias{amplifyTrans}
\title{Write out mediator-based trans-eSNP effects}
\usage{
amplifyTrans(whichMed, medTrainList, lmCaretObj)
}
\arguments{
\item{geneInt}{character, identifier for gene of interest}

\item{qtlFull}{data frame, mediator-gene qtl results}

\item{numMed}{numeric, user-defined maximum number of correlated mediators}
}
\value{
a vector of correlated biomarkers to be treated as mediators
}
\description{
The function takes in the SNP-weights for mediators and
writes out the amplified SNP-weights based on
}
