% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burdenTest.R
\name{burdenTest}
\alias{burdenTest}
\title{Compute weighted burden test}
\usage{
burdenTest(
  wgt,
  snps,
  sumStats,
  snpAnnot = NULL,
  onlyCis = F,
  Z = NULL,
  beta = NULL,
  se = NULL,
  chr,
  pos,
  ref,
  pval,
  R2cutoff,
  alpha,
  nperms = 1000
)
}
\arguments{
\item{wgt}{character, name of expression model .RData file}

\item{snps}{data.frame, reference panel SNPs}

\item{sumStats}{data frame, GWAS summary statistics}

\item{snpAnnot}{data.frame, SNP REF/ALT annotations}

\item{onlyCis}{logical, T/F to consider only cis-component}

\item{beta}{character, colnames in sumStats that keeps the effect sizes}

\item{se}{character, colnames in sumStats that keeps the standard errors}

\item{chr}{character, colnames in sumStats that keeps the chromosome}

\item{pos}{character, colnames in sumStats that keeps the position}

\item{ref}{character, colnames in sumStats that keeps the reference allele}

\item{R2cutoff}{numeric, predictive R2 cutoff}

\item{alpha}{numeric, P-value threshold for permutation testing}

\item{nperms}{numeric, number of permutations}
}
\value{
list of results for burden and permutation tests
}
\description{
The function takes in a gene expression model in MOSTWAS form
and GWAS summary statistics and
carries out the weighted burden Z-test for a trait
}
