% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeTME.R
\name{computeTME}
\alias{computeTME}
\title{Compute total mediation effect}
\usage{
computeTME(snp, expression, mediators, covs, indices)
}
\arguments{
\item{snp}{vector, SNP of interest}

\item{expression}{vector, gene expression of interest}

\item{mediators}{data frame, mediators of interest}

\item{covs}{data frame, covariates}

\item{indices}{blank, index for boot package}
}
\value{
estimate of TME
}
\description{
The function takes in a SNP-mediator set-gene triplet
and computes the total mediation effect with boots structure
}
