% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertToDosage.R
\name{convertToDosage}
\alias{convertToDosage}
\title{Train and predict gene's predictive model with mediators}
\usage{
convertToDosage(format = "plink", fileName, doseFile, locsFile,
  annotFile, append = F)
}
\arguments{
\item{format}{character, 'plink' or 'vcf'}

\item{fileName}{character, prefix only for plink or vcf files}

\item{doseFile}{character, output file name for dosage file}

\item{locsFile}{character, output file name for SNP location file}

\item{annotFile}{character, output file name for SNP annotation file}

\item{append}{logical, T/F if we are appending to existing files for large plink/vcf files}
}
\value{
final model for gene along with CV R2 and predicted values
}
\description{
The function trains a predictive model of a given gene using top mediators
as fixed effects and assesses in-sample performance with cross-validation.
}
