% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVFitMediator.R
\name{CVFitMediator}
\alias{CVFitMediator}
\title{Fit a single split in cross-validation for a given biomarker}
\usage{
CVFitMediator(
  foldNum,
  parts.train,
  parts.test,
  pheno,
  snpCur,
  parallel = T,
  fixedEffects = NULL
)
}
\arguments{
\item{foldNum}{integer, which fold of the k splits}

\item{parts.train}{list, list of training indices}

\item{parts.test}{list, list of test indices}

\item{pheno}{vector, vector of standardized intensities of biomarker}

\item{snpCur}{matrix, matrix of cis-genotypes to biomarker}

\item{parallel}{logical, run in parallel?}
}
\value{
data frame of elastic net, lasso, and LMM based predictions
}
\description{
The function trains a predictive model on the given training split and then
predicts the genetically regulated intensity of it in the test split.
}
