% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DePMA.R
\name{DePMA}
\alias{DePMA}
\title{Train and predict gene's predictive model with mediators using DePMA}
\usage{
DePMA(
  geneInt,
  snpObj,
  mediator,
  medLocs,
  covariates,
  cisDist = 5e+05,
  qtlTra,
  qtMed,
  h2Pcutoff,
  dimNumeric,
  verbose,
  seed,
  sobel = F,
  nperms = 1000,
  k,
  parallel,
  parType = "no",
  prune,
  ldThresh = 0.5,
  cores,
  qtlTra_parts,
  qtMed_parts,
  modelDir,
  tempFolder,
  R2Cutoff
)
}
\arguments{
\item{geneInt}{character, identifier for gene of interest}

\item{snpObj}{binsnp object, SNP dosages}

\item{mediator}{data frame, mediator intensities}

\item{medLocs}{data frame, MatrixEQTL locations for mediators}

\item{covariates}{data frame, covariates}

\item{h2Pcutoff}{numeric, P-value cutoff for heritability}

\item{dimNumeric}{numeric, number of numeric covariates}

\item{verbose}{logical, output everything}

\item{seed}{integer, random seed for splitting}

\item{sobel}{logical, Sobel asymptotic test T/F}

\item{nperms}{numeric, number of permutations}

\item{k}{integer, number of training-test splits}

\item{parallel}{logical, TRUE/FALSE to run glmnet in parallel}

\item{parType}{character, parallelization type for boots}

\item{prune}{logical, TRUE/FALSE to LD prune the genotypes}

\item{ldThresh}{numeric, LD threshold for PLINK pruning}

\item{cores}{integer, number of parallel cores}

\item{qtlTra_parts}{character vector, files for SNP to gene distal-eqtls}

\item{qtMed_parts}{character vector, files for SNP to mediators local-eqtls}

\item{modelDir}{character, directory for saving models}

\item{tempFolder}{character, directory of saving snp backing files}

\item{R2Cutoff}{numeric, cutoff for model R2}

\item{qtlFull}{data frame, all QTLs (cis and trans) between mediators and genes}

\item{numMed}{integer, number of top mediators to include}
}
\value{
final model for gene along with CV R2 and predicted values
}
\description{
The function trains a predictive model of a given gene using top mediators
as fixed effects and assesses in-sample performance with cross-validation.
}
