% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eQTL_MOSTWAS.R
\name{eQTL_MOSTWAS}
\alias{eQTL_MOSTWAS}
\title{LD prune a genotype matrix}
\usage{
eQTL_MOSTWAS(
  SNP_file_name,
  snps_location_file_name,
  SNP_file_name_dis = NULL,
  expression_file_name,
  gene_location_file_name,
  mediators_file_name,
  meds_location_file_name,
  covariates_file_name,
  output_file_name_loc_qtl,
  output_file_name_dis_qtl,
  output_file_name_loc_med = NULL,
  output_file_name_dis_med = NULL,
  p_loc_qtl = 1e-06,
  p_dis_qtl = 1e-06,
  FDRcut = 0.05,
  useModel = modelLINEAR,
  DePMA = F
)
}
\arguments{
\item{SNP_file_name}{character, file name for SNP dosages in MatrixEQTL format}

\item{snps_location_file_name}{character, file name for SNP locations in MatrixEQTL format}

\item{SNP_file_name_dis}{character, file name for SNP distal dosages}

\item{expression_file_name}{character, file name for gene expressions}

\item{gene_location_file_name}{character, file name for gene locations}

\item{mediators_file_name}{character, file name for mediator intensities}

\item{meds_location_file_name}{character, file name for mediator locations as genes}

\item{output_file_name_loc_qtl}{character, file name for first local QTLs}

\item{output_file_name_dis_qtl}{character, file name for first distal QTLs}

\item{output_file_name_loc_med}{character, file name for second local QTLs}

\item{output_file_name_dis_med}{character, file name for second distal QTLs}

\item{p_loc_qtl}{numeric, P-value threshold for first QTLs for DePMA}

\item{p_dis_qtl}{numeric, P-value threshold for first QTLs for DePMA}

\item{FDRcut}{numeric, FDR-adjusted P-value threshold for second QTLs for DePMA}

\item{useModel}{MatrixEQTL model type}

\item{DePMA}{logical, T/F if this analysis is for DePMA}

\item{covariates_files_name}{character, file name for covariaties}
}
\value{
list of pruned genotype matrix, vector of SNP names, and location data frame
}
\description{
The function prunes SNPs that are in LD using PLINK
}
