% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateHeritability.R
\name{estimateHeritability}
\alias{estimateHeritability}
\title{Estimate germline heritability of a biomarker}
\usage{
estimateHeritability(biomInt, snps, pheno, snpLocs, mediator, medLocs,
  covariates, fileName, dimNumeric, numMed = 5, cisDist = 5e+05,
  needMed = T, prune = F, medList, verbose = F, windowSize,
  numSNPShift, ldThresh, ldScrRegion = 200, LDMS = F, supplySNP = F,
  snpAnnot = NULL)
}
\arguments{
\item{biomInt}{character, identifier for biomarker of interest}

\item{snps}{data frame, SNP dosages}

\item{snpLocs}{data frame, MatrixEQTL locations for SNPs}

\item{mediator}{data frame, mediator intensities}

\item{medLocs}{data frame, MatrixEQTL locations for mediators}

\item{covariates}{data frame, covariates}

\item{fileName}{character, throw away name for PLINK files}

\item{dimNumeric}{integer, number of continuous covariates}

\item{numMed}{integer, number of top mediators to include}

\item{cisDist}{numeric, base pair window of cis definition}

\item{needMed}{logical, T/F if the biomarker of interest has associated mediators}

\item{medList}{character vector,  vector of mediator names}

\item{verbose}{logical, T/F for detailed output}

\item{windowSize}{integer, window size for PLINK pruning}

\item{numSNPShift}{integer, shifting window from PLINK pruning}

\item{ldThresh}{numeric, maximum threshold for linkage disequilibrium}

\item{ldScrRegion}{numeric, number of SNPs per bin in GCTA-LDMS}

\item{LDMS}{logical, T/F to run GCTA-LDMS}

\item{supplySNP}{logical, T/F to run GCTA on the full provided snps}

\item{numMed}{integer, maximum number of mediators to find}
}
\value{
heritability estimate and associated LRT P-value
}
\description{
The function invokves GCTA in the system and estimates the germline
heritability of a biomarker given cis-regions around the biomarker
and any associated mediators of interest.
}
