% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatherMediators.R
\name{gatherMediators}
\alias{gatherMediators}
\title{Extract mediating biomarkers for a given gene}
\usage{
gatherMediators(geneInt, qtlFull, numMed)
}
\arguments{
\item{geneInt}{character, identifier for gene of interest}

\item{qtlFull}{data frame, mediator-gene qtl results}

\item{numMed}{numeric, user-defined maximum number of correlated mediators}
}
\value{
a vector of correlated biomarkers to be treated as mediators
}
\description{
The function runs aggregates top correlated biomarkers
to a given gene of interest and exports the list
}
