% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCisGenotypes.R
\name{getCisGenotypes}
\alias{getCisGenotypes}
\title{Gather the cis-genotypes for a given biomarker}
\usage{
getCisGenotypes(biomInt, locs, snps, snpLocs, cisDist = 5e+05)
}
\arguments{
\item{biomInt}{character, identifier for biomarker of interest}

\item{locs}{data frame, location file for the biomarkers}

\item{snps}{data frame, SNP dosages}

\item{snpLocs}{data frame, MatrixEQTL locations for SNPs}

\item{cisDist}{numeric, window for cis distance}
}
\value{
list with matrix of cis-genotypes, vector of SNP names, and data frame of SNP locations
}
\description{
The function takes in a biomarker of interest (mediator or gene)
and gather the cis-genotypes in a user-defined window around
the biomarker.
}
