% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDprune.R
\name{LDprune}
\alias{LDprune}
\title{LD prune a genotype matrix}
\usage{
LDprune(W, snpList, snpLocs, fileName, windowSize, numSNPShift, ldThresh,
  verbose = T, snpAnnot = NULL)
}
\arguments{
\item{W}{matrix, dosage of SNPs}

\item{snpList}{character vector, vector of SNPs in W}

\item{snpLocs}{data frame, SNP locations in MatrixEQTL form}

\item{fileName}{character, name for PLINK intermediate files}

\item{windowSize}{integer, window size for PLINK pruning}

\item{numSNPShift}{integer, shifting window for PLINK pruning}

\item{ldThresh}{numeric, LD threshold for PLINK pruning}
}
\value{
list of pruned genotype matrix, vector of SNP names, and location data frame
}
\description{
The function prunes SNPs that are in LD using PLINK
}
