% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeFinalModels.R
\name{mergeFinalModels}
\alias{mergeFinalModels}
\title{Output final MOSTWAS models}
\usage{
mergeFinalModels(
  MeTWASFolder,
  DePMAFolder,
  outFolder,
  sumFile,
  R2Cutoff = 0.01,
  h2PCutoff = 0.05
)
}
\arguments{
\item{MeTWASFolder}{character, folder name for MeTWAS models}

\item{DePMAFolder}{character, folder name for DePMA models}

\item{outFolder}{character, folder name for final models folder}
}
\value{
no return object, creates a folder with final MOSTWAS models
}
\description{
The function computes the adjusted R2 between two numeric vectors
}
