% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeTWAS.R
\name{MeTWAS}
\alias{MeTWAS}
\title{Train and predict gene's predictive model with mediators using MeTWAS}
\usage{
MeTWAS(
  geneInt,
  snpObj,
  mediator,
  medLocs,
  covariates,
  dimNumeric,
  qtlFull,
  h2Pcutoff = 0.1,
  numMed = 5,
  seed = 1218,
  k = 5,
  cisDist = 1e+06,
  parallel = T,
  prune = F,
  ldThresh = 0.5,
  cores,
  verbose = T,
  R2Cutoff = 0.01,
  modelDir,
  tempFolder
)
}
\arguments{
\item{geneInt}{character, identifier for gene of interest}

\item{snpObj}{binsnp object, SNP dosages}

\item{mediator}{data frame, mediator intensities}

\item{medLocs}{data frame, MatrixEQTL locations for mediators}

\item{covariates}{data frame, covariates}

\item{dimNumeric}{numeric, number of numeric covariates}

\item{qtlFull}{data frame, all QTLs (cis and trans) between mediators and genes}

\item{h2Pcutoff}{numeric, P-value cutoff for heritability}

\item{numMed}{integer, number of top mediators to include}

\item{seed}{integer, random seed for splitting}

\item{k}{integer, number of training-test splits}

\item{parallel}{logical, TRUE/FALSE to run glmnet in parallel}

\item{prune}{logical, TRUE/FALSE to LD prune the genotypes}

\item{ldThresh}{numeric, LD threshold for PLINK pruning}

\item{cores}{integer, number of parallel cores}

\item{verbose}{logical, output everything}

\item{R2Cutoff}{numeric, cutoff for model R2}

\item{modelDir}{character, directory for saving models}

\item{tempFolder}{character, directory of saving snp backing files}
}
\value{
final model for gene along with CV R2 and predicted values
}
\description{
The function trains a predictive model of a given gene using top mediators
as fixed effects and assesses in-sample performance with cross-validation.
}
