% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regress.R
\name{regress}
\alias{regress}
\title{Generate summary statistics for genotype-trait association}
\usage{
regress(Z, pheno)
}
\arguments{
\item{Z}{matrix, simulated genotypes}

\item{pheno}{vector, trait}
}
\value{
data frame of effect sizes, standard errors and P-values of associations
}
\description{
The function takes in simulated genotypes and trait
and outputs summary statistics for associations.
}
