% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepParallel.R
\name{RepParallel}
\alias{RepParallel}
\title{Parallel implementation of replicate()}
\usage{
RepParallel(n, expr, simplify = "array", mc.cores = 1)
}
\arguments{
\item{n}{number of replications}

\item{expr}{expression}

\item{simplify}{character, simplify to array}
}
\value{
array of replications
}
\description{
The function parallelizes replicate(), as in dbframe-R-library
}
