% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simBeta.R
\name{simBeta}
\alias{simBeta}
\title{Simulate eQTL effect sizes}
\usage{
simBeta(p.causal, eqtl_h2, n_snps)
}
\arguments{
\item{p.causal}{numeric, proportion of causal eQTLs}

\item{eqtl_h2}{numeric, heritability of expression from eQTLs}

\item{n_snps}{integer, number of eQTLs}
}
\value{
vector of effect sizes
}
\description{
The function takes in a proportion
of causal eQTLs and total heritability of eQTLs
and generates n_snps eQTL effect sizes
}
