% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simEQTL.R
\name{simEQTL}
\alias{simEQTL}
\title{Simulate eQTL reference panel}
\usage{
simEQTL(geno.loc, geno.dis, nqtl, b_qtls.loc, b_qtls.dis, eqtl_h2.loc,
  eqtl_h2.dis, numMed)
}
\arguments{
\item{geno.loc}{data.frame, reference genotypes for local SNPs}

\item{geno.dis}{data.frame, reference genotypes for distal SNPs}

\item{nqtl}{integer, sample size}

\item{b_qtls.loc}{vector, local eQTLs effect sizes}

\item{b_qtls.dis}{vector, distal eQTLs effect sizes}

\item{eqtl_h2.loc}{numeric, local heritability of expression}

\item{eqtl_h2.dis}{numeric, distal heritability of expression}

\item{numMed}{integer, number of mediators}
}
\value{
list of matrices of local and distal simulated genotypes, expression,
mediator intensities, vector of final distal eQTL effect sizes
}
\description{
The function creates a eQTL reference panel for MOSTWAS
methods with genotypes, expression, mediator intensities, and QTLs.
}
