% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simGeno.R
\name{simGeno}
\alias{simGeno}
\title{Simulate genotypes from reference data}
\usage{
simGeno(geno, n)
}
\arguments{
\item{geno}{data.frame, reference panel SNPs}

\item{n}{integer, samples size}
}
\value{
matrix of simulated genotypes
}
\description{
The function takes in a reference genotype data
and generates simulated genotypes from the LD structure.
Modified from Nicholas Mancuso's twas_sim.
}
