% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simTrait.R
\name{simTrait}
\alias{simTrait}
\title{Simulate trait from latent genetic values}
\usage{
simTrait(g, h2g)
}
\arguments{
\item{g}{vector, latent genetic values}

\item{h2g}{numeric, heritability of trait}
}
\value{
vector of trait
}
\description{
The function takes in a vector of latent genetic
values and simulates a trait with controlled heritable variance
explained
}
