% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sobelTest.R
\name{sobelTest}
\alias{sobelTest}
\title{Compute total mediation effect}
\usage{
sobelTest(snp, expression, mediators, covs)
}
\arguments{
\item{snp}{vector, SNP of interest}

\item{expression}{vector, gene expression of interest}

\item{mediators}{data frame, mediators of interest}

\item{covs}{data frame, covariates}
}
\value{
estimate of TME and Sobel test P-value
}
\description{
The function takes in a SNP-mediator set-gene triplet
and computes the total mediation effect and tests
the total mediator effect via a Sobel test (1982)
}
