% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testTME.R
\name{testTME}
\alias{testTME}
\title{Run TME tests through a data.frame of mediators}
\usage{
testTME(
  i,
  mediator,
  qtMed,
  transSNPs,
  pheno,
  sobel = F,
  nperms = 1000,
  cores,
  covariates,
  parallel = "no"
)
}
\arguments{
\item{i}{integer, row of data.frame}

\item{mediator}{data.frame, mediator intensities}

\item{qtMed}{data.frame, mediator QTL results}

\item{transSNPs}{vector, character vector of distal-eSNPs}

\item{sobel}{logical, T/F to use Sobel test}

\item{nperms}{integer, number of permutations}

\item{cores}{integer, number of parallel ocres}

\item{covariates}{data.frame, covariates}

\item{parallel}{character, boots input for parallelization}
}
\value{
list with TME and P-value
}
\description{
The function runs Sobel or permutation testing
}
