% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainDeP.R
\name{trainDeP}
\alias{trainDeP}
\title{Train and predict gene's predictive model with mediators}
\usage{
trainDeP(geneInt, snps, snpLocs, mediator, medLocs, covariates,
  cisDist = 5e+05, qtlTra, qtMed, h2Pcutoff, dimNumeric, verbose, seed,
  sobel = F, nperms = 1000, k, parallel, parType = "no", prune,
  windowSize = 50, numSNPShift = 5, ldThresh = 0.5, cores,
  qtlTra_parts, qtMed_parts, modelDir, snpAnnot = NULL)
}
\arguments{
\item{geneInt}{character, identifier for gene of interest}

\item{snps}{data frame, SNP dosages}

\item{snpLocs}{data frame, MatrixEQTL locations for SNPs}

\item{mediator}{data frame, mediator intensities}

\item{medLocs}{data frame, MatrixEQTL locations for mediators}

\item{covariates}{data frame, covariates}

\item{h2Pcutoff}{numeric, P-value cutoff for heritability}

\item{seed}{integer, random seed for splitting}

\item{k}{integer, number of training-test splits}

\item{parallel}{logical, TRUE/FALSE to run glmnet in parallel}

\item{prune}{logical, TRUE/FALSE to LD prune the genotypes}

\item{windowSize}{integer, window size for PLINK pruning}

\item{numSNPShift}{integer, shifting window for PLINK pruning}

\item{ldThresh}{numeric, LD threshold for PLINK pruning}

\item{cores}{integer, number of parallel cores}

\item{qtlFull}{data frame, all QTLs (cis and trans) between mediators and genes}

\item{numMed}{integer, number of top mediators to include}

\item{outputAll}{logical, include mediator information}
}
\value{
final model for gene along with CV R2 and predicted values
}
\description{
The function trains a predictive model of a given gene using top mediators
as fixed effects and assesses in-sample performance with cross-validation.
}
