% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainLocalModel.R
\name{trainLocalModel}
\alias{trainLocalModel}
\title{Train a local-only model for a phenotype}
\usage{
trainLocalModel(
  phenoInt,
  midSNP,
  mediator,
  medLocs,
  covariates,
  cisDist = 5e+05,
  nfolds = 5,
  seed = 1218
)
}
\arguments{
\item{phenoInt}{character, phenotype name}

\item{midSNP}{snpObject, snp Object in the bigsnpr format}

\item{mediator}{data frame, mediator intensities}

\item{medLocs}{data frame, mediator locations}

\item{covariates}{data frame, covariates in MatrixEQTL format}

\item{cisDist}{numeric, local window, default .5 Mb}

\item{nfolds}{integer, number of folds}

\item{seed}{numeric, random seed}
}
\value{
list with Model, predicted values, and CV-R2
}
\description{
The function runs FUSION-type local-only modeling for a phenotype
}
