% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainMediator.R
\name{trainMediator}
\alias{trainMediator}
\title{Train and predict mediator predictive model}
\usage{
trainMediator(
  medInt,
  pheno = NULL,
  mediator,
  medLocs,
  snps,
  snpLocs,
  covariates,
  seed,
  k,
  cisDist = 5e+05,
  prune = T,
  windowSize = 50,
  numSNPShift = 5,
  ldThresh = 0.5,
  snpAnnot = NULL
)
}
\arguments{
\item{medInt}{character, identifier for mediator of interest}

\item{mediator}{data frame, mediator intensities}

\item{medLocs}{data frame, MatrixEQTL locations for mediators}

\item{snps}{data frame, SNP dosages}

\item{snpLocs}{data frame, MatrixEQTL locations for SNPs}

\item{covariates}{data frame, covariates}

\item{seed}{integer, random seed for splitting}

\item{k}{integer, number of training-test splits}

\item{prune}{logical, TRUE/FALSE to LD prune the genotypes}

\item{windowSize}{integer, window size for PLINK pruning}

\item{numSNPShift}{integer, shifting window for PLINK pruning}

\item{ldThresh}{numeric, LD threshold for PLINK pruning}

\item{parallel}{logical, TRUE/FALSE to run glmnet in parallel}

\item{cores}{integer, number of parallel cores}
}
\value{
final model for mediator along with CV R2 and predicted values
}
\description{
The function trains a predictive model of a given mediator and then
predicts the genetically regulated intensity of it in the training set via
cross-validation.
}
