% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainSNPPheno.R
\name{trainSNPPheno}
\alias{trainSNPPheno}
\title{Train a local-only model for a phenotype}
\usage{
trainSNPPheno(
  pheno,
  snpCur,
  snpList,
  thisSNP,
  fileName,
  prune = T,
  windowSize = 50,
  numSNPShift = 5,
  ldThresh = 0.5,
  verbose = T,
  snpAnnot = NULL
)
}
\arguments{
\item{pheno}{vector, phenotypes}

\item{snpCur}{matrix, SNP dosages}

\item{snpList}{vector, vector of SNP identifiers}

\item{thisSNP}{data.frame, SNP locations}

\item{fileName}{character, prefix for LDprune}

\item{prune}{logical, T/F to LD prune}

\item{windowSize}{numeric, PLINK LD prune window size}

\item{numSNPShift}{integer, PLINK LD shifting interval}

\item{ldThresh}{numeric, LD threshold}

\item{verbose}{logical, T/F for verbosity}

\item{snpAnnot}{data.frame, SNP annotations}
}
\value{
list with TME and P-value
}
\description{
The function runs FUSION-type local-only modeling for a phenotype
}
